Renamed file and module liststyle -> listproperties
