+++ packages/ckeditor5-list/ckeditor5-metadata.json
@@ -75,7 +75,7 @@
 			"className": "ListProperties",
 			"description": "Enables styling the list item markers for both ordered and unordered lists. You can choose various types of numerals and letters or visual markers to use with these lists.",
 			"docs": "features/lists/lists.html#list-styles",
-			"path": "src/liststyle.js",
+			"path": "src/listproperties.js",
 			"requires": [
 				"List"
 			],
