+++ packages/ckeditor5-list/docs/_snippets/features/lists-source.js
@@ -6,7 +6,7 @@
 /* globals window */
 
 import ClassicEditor from '@ckeditor/ckeditor5-build-classic/src/ckeditor';
-import ListProperties from '@ckeditor/ckeditor5-list/src/liststyle';
+import ListProperties from '@ckeditor/ckeditor5-list/src/listproperties';
 
 window.ClassicEditor = ClassicEditor;
 window.ListProperties = ListProperties;
