+++ packages/ckeditor5-list/docs/features/lists.md
@@ -37,7 +37,7 @@ A Markdown code provided by the {@link features/autoformat autoformatting featur
 
 ## List styles
 
-The {@link module:list/liststyle~ListProperties list style} feature introduces some more styles for the list item markers. When enabled, it adds 3 styles for unordered lists and 6 styles for ordered lists to choose from. The styles can be changed via the dropdown that opens when you click the arrow next to the appropriate list button in the toolbar.
+The {@link module:list/listproperties~ListProperties list style} feature introduces some more styles for the list item markers. When enabled, it adds 3 styles for unordered lists and 6 styles for ordered lists to choose from. The styles can be changed via the dropdown that opens when you click the arrow next to the appropriate list button in the toolbar.
 
 ### Demo
 
@@ -63,7 +63,7 @@ npm install --save @ckeditor/ckeditor5-list
 Then add the `ListProperties` plugin to your plugin list and the toolbar configuration:
 
 ```js
-import ListProperties from '@ckeditor/ckeditor5-list/src/liststyle';
+import ListProperties from '@ckeditor/ckeditor5-list/src/listproperties';
 
 ClassicEditor
 	.create( document.querySelector( '#editor' ), {
@@ -79,7 +79,7 @@ ClassicEditor
 </info-box>
 
 <info-box warning>
-	The {@link module:list/liststyle~ListProperties} feature overrides UI button implementations from the {@link module:list/listui~ListUI}.
+	The {@link module:list/listproperties~ListProperties} feature overrides UI button implementations from the {@link module:list/listui~ListUI}.
 </info-box>
 
 ## List indentation
@@ -97,7 +97,7 @@ The {@link module:list/list~List} plugin registers:
 * The `'numberedList'` UI button.
 * The `'bulletedList'` UI button.
 
-The {@link module:list/liststyle~ListProperties} plugin registers:
+The {@link module:list/listproperties~ListProperties} plugin registers:
 
 * The {@link module:list/liststylecommand~ListStyleCommand `'listStyle'`} command that accepts a `type` of the list style to set.
     ```js
