+++ packages/ckeditor5-list/src/index.js
@@ -10,7 +10,7 @@
 export { default as List } from './list';
 export { default as ListEditing } from './listediting';
 export { default as ListUI } from './listui';
-export { default as ListProperties } from './liststyle';
+export { default as ListProperties } from './listproperties';
 export { default as ListPropertiesEditing } from './listpropertiesediting';
 export { default as ListPropertiesUI } from './listpropertiesui';
 export { default as TodoList } from './todolist';
