+++ packages/ckeditor5-list/src/listreversedcommand.js
@@ -12,7 +12,7 @@ import { getSelectedListItems } from './utils';
 
 /**
  * The list reversed command. It changes `listReversed` attribute of the selected list items.
- * It is used by the {@link module:list/liststyle~ListProperties list style feature}.
+ * It is used by the {@link module:list/listproperties~ListProperties list style feature}.
  *
  * @extends module:core/command~Command
  */
