+++ packages/ckeditor5-list/src/liststartcommand.js
@@ -12,7 +12,7 @@ import { getSelectedListItems } from './utils';
 
 /**
  * The list start index command. It changes `listStart` attribute of the selected list items.
- * It is used by the {@link module:list/liststyle~ListProperties list style feature}.
+ * It is used by the {@link module:list/listproperties~ListProperties list style feature}.
  *
  * @extends module:core/command~Command
  */
